AIHeader -> HeroHeader
- Since this will also be reused as the editorial header, makes sense
to put this in one place to be shared.
