+++ src/components/editorials/EditorialParts.js
@@ -1,6 +1,7 @@
-// @flow
 /* eslint-disable react/no-danger */
 /* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable javascript/flow */
+
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router'
